_deprecated/entity-cli_v1/assets/core/xxxxeclixxxx/infra/in-memory/xxxxeclixxxx-inmemory.repository.ts
import { InMemoryRepository } from '../../../shared';
import { IXxxxeclixxxx } from '../../application';
import { Xxxxeclixxxx, XxxxeclixxxxRepositoryInterface } from '../../domain';

export class XxxxeclixxxxInMemoryRepository
  extends InMemoryRepository<
    IXxxxeclixxxx,
    Xxxxeclixxxx,
    XxxxeclixxxxRepositoryInterface.XxxxeclixxxxFields
  >
  implements XxxxeclixxxxRepositoryInterface.Repository
{
  sortableFields: XxxxeclixxxxRepositoryInterface.XxxxeclixxxxFields[] = [
    'name',
    'created_at',
  ];
  searchableFields: XxxxeclixxxxRepositoryInterface.XxxxeclixxxxFields[] = [
    'name',
  ];
  filterableFields: XxxxeclixxxxRepositoryInterface.XxxxeclixxxxFields[] = [
    'id',
    'name',
    'status',
  ];
}
